+++ CHANGELOG.md
@@ -5,4 +5,5 @@
 - CHANGELOG.md (this file)
 - This tag as starting point for further changes
 - Frontpage now reads this file to make users aware of changes.
+- detailed-move component to tidy up detailed grid item
 - Happy birthday 🥝!